Move formatDate and date constants under lib-common
