+++ frontend/src/citizen-frontend/applications/editor/verification/BasicsSection.tsx
@@ -7,9 +7,9 @@ import React from 'react'
 import { useTranslation } from '../../../localization'
 import { H2, Label } from 'lib-components/typography'
 import ListGrid from 'lib-components/layout/ListGrid'
-import { DATE_FORMAT_DATE_DEFAULT, formatDate } from '../../../util'
 import { ApplicationDataGridLabelWidth } from '../../../applications/editor/verification/const'
 import { ApplicationDetails } from 'lib-common/api-types/application/ApplicationDetails'
+import { formatDate } from 'lib-common/date'
 
 type BasicsSectionProps = {
   application: ApplicationDetails
@@ -34,17 +34,13 @@ export default React.memo(function BasicsSection({
         columnGap="L"
       >
         <Label>{t.applications.editor.verification.basics.created}</Label>
-        <span>
-          {formatDate(application.createdDate, DATE_FORMAT_DATE_DEFAULT)}
-        </span>
+        <span>{formatDate(application.createdDate)}</span>
 
-        {formatDate(application.modifiedDate, DATE_FORMAT_DATE_DEFAULT) !==
-          formatDate(application.createdDate, DATE_FORMAT_DATE_DEFAULT) && (
+        {formatDate(application.modifiedDate) !==
+          formatDate(application.createdDate) && (
           <>
             <Label>{t.applications.editor.verification.basics.modified}</Label>
-            <span>
-              {formatDate(application.modifiedDate, DATE_FORMAT_DATE_DEFAULT)}
-            </span>
+            <span>{formatDate(application.modifiedDate)}</span>
           </>
         )}
       </ListGrid>
