+++ frontend/src/citizen-frontend/messages/ThreadListItem.tsx
@@ -10,8 +10,8 @@ import colors from 'lib-customizations/common'
 import React from 'react'
 import styled from 'styled-components'
 import { useTranslation } from '../localization'
-import { formatDate } from '../util'
 import { MessageTypeChip } from './MessageTypeChip'
+import { formatDate } from 'lib-common/date'
 
 interface Props {
   thread: MessageThread
