+++ frontend/src/employee-frontend/components/application-page/ApplicationNoteBox.tsx
@@ -7,9 +7,8 @@ import { ApplicationNote } from '../../types/application'
 import styled from 'styled-components'
 import { greyscale } from 'lib-customizations/common'
 import { Label } from 'lib-components/typography'
-import { formatDate } from '../../utils/date'
+import { DATE_FORMAT_DATE_TIME, formatDate } from 'lib-common/date'
 import { formatParagraphs } from '../../utils/html-utils'
-import { DATE_FORMAT_DATE_TIME } from '../../constants'
 import { defaultMargins, Gap } from 'lib-components/white-space'
 import { TextArea } from 'lib-components/atoms/form/InputField'
 import { addSeconds, isAfter } from 'date-fns'
