+++ frontend/src/employee-frontend/components/fee-decisions/FeeDecisions.tsx
@@ -23,7 +23,7 @@ import { useTranslation } from '../../state/i18n'
 import { FeeDecisionSummary } from '../../types/invoicing'
 import { SearchOrder } from '../../types'
 import { Result } from 'lib-common/api'
-import { formatDate } from '../../utils/date'
+import { formatDate } from 'lib-common/date'
 import { formatCents } from '../../utils/money'
 import { SortByFeeDecisions } from '../../api/invoicing'
 import Pagination from 'lib-components/Pagination'
