+++ frontend/src/employee-frontend/components/messages/SingleThreadView.tsx
@@ -17,10 +17,9 @@ import colors from 'lib-customizations/common'
 import { faAngleLeft } from 'lib-icons'
 import React, { useCallback, useContext, useMemo } from 'react'
 import styled from 'styled-components'
-import { DATE_FORMAT_DATE_TIME } from '../../constants'
 import { useTranslation } from '../../state/i18n'
 import { UUID } from '../../types'
-import { formatDate } from '../../utils/date'
+import { DATE_FORMAT_DATE_TIME, formatDate } from 'lib-common/date'
 import { MessageContext } from './MessageContext'
 import { MessageTypeChip } from './MessageTypeChip'
