+++ frontend/src/employee-frontend/constants.ts
@@ -6,20 +6,6 @@ import LocalDate from 'lib-common/local-date'
 import { UUID } from './types'
 import { AssistanceBasis } from './types/child'
 
-export const DATE_FORMAT_DEFAULT = 'dd.MM.yyyy'
-export const DATE_FORMAT_DATE_TIME = 'dd.MM.yyyy HH:mm'
-export const DATE_FORMAT_NO_YEAR = 'dd.MM.'
-export const DATE_FORMAT_ISO = 'yyyy-MM-dd'
-export const DATE_FORMAT_TIME_ONLY = 'HH:mm'
-export const DATE_FORMATS_PARSED = [
-  'dd.MM.yyyy',
-  'dd.MM.yy',
-  'd.M.yyyy',
-  'd.M.yy',
-  'ddMMyyyy',
-  'ddMMyy'
-]
-
 export const EVAKA_START = LocalDate.of(2020, 3, 1)
 
 export const MAX_DATE = LocalDate.of(2099, 12, 31)
