+++ frontend/src/employee-frontend/utils/date.spec.ts
@@ -3,7 +3,7 @@
 // SPDX-License-Identifier: LGPL-2.1-or-later
 
 import LocalDate from 'lib-common/local-date'
-import { rangesOverlap, formatDate, autoComplete } from './date'
+import { rangesOverlap, autoComplete } from './date'
 
 describe('utils/date', () => {
   describe('rangesOverlap', () => {
@@ -174,16 +174,6 @@ describe('utils/date', () => {
         }
       )
       expect(overlap).toBe(false)
-    })
-  })
-
-  describe('formatDate', () => {
-    it('should format valid date correctly', () => {
-      expect(formatDate(new Date('2019-01-01'))).toBe('01.01.2019')
-    })
-
-    it('should format undefined to empty string', () => {
-      expect(formatDate(undefined)).toBe('')
     })
   })
